fix(open-file): move openFileTool to inline-tools, decouple from recording
`openFileTool` lived in `src/recording-tools.ts` since #274 with
recording-flavored helpers (`findRecording()` fallback, `sutando-recording`
substring check, `/tmp/sutando-playback-path` write, `demoStateRef.value
= 'idle'` reset, ffprobe duration probe) coupled into the function body.
PR #353 introduced most of those; PR #392 cleaned the worst (made
`find_recording=true` opt-in), but kept the helpers as deliberate residuals.
This commit removes them entirely and moves the function to a generic
home.

Changes:

1. **Move `openFileTool` to `src/inline-tools.ts`.** Required `path` only.
   Stripped: `find_recording`, `findRecording()` fallback,
   `sutando-recording` substring write, `demoStateRef` reset,
   `recPath` rename → `filePath`, "play_video" instruction string,
   ffprobe duration extraction.

2. **Keep `fullscreen=true` flow in `openFileTool`** (per Chi's confirmation
   in #dev — fullscreen-on-open is general video-fullscreen UX, not
   recording-specific). The QuickTime force-open + present-mode
   AppleScript + Zoom-survives keystroke routing all stay. The runtime
   guard ("don't fullscreen=true when QT already has docs") stays.

3. **Re-export hygiene:**
   - `inline-tools.ts` line-17 re-export and line-18 import: dropped
     `openFileTool` (now defined locally in this file).
   - `browser-tools.ts` recording-tools re-export block: dropped
     `openFileTool`.
   - `recording-tools.ts`: removed definition + the misleading
     "polymorphic playRecordingTool" comment block above it. One-line
     marker pointing to the new location.

4. **Recording-flavored side effects relocated to recording-tools:**
   - `/tmp/sutando-playback-path` write moved to `screenRecordTool`
     stop handler AND its auto-stop `setTimeout` callback. Recording
     tool now stamps the playback hint itself when the recording
     finishes — no model courier through `open_file` needed.
   - ffprobe duration extraction moved to `screenRecordTool` stop
     result (`duration_seconds` field).
   - Stop handler instruction updated to mention both
     `open_file(path)` and `open_file(path, fullscreen=true)` so the
     model knows how to invoke each path of the demo flow.

README demo flow (`screen_record → open_file(path) → play_video`)
continues to work end-to-end: `screen_record stop` writes
`/tmp/sutando-playback-path` proactively, and `play_video`'s existing
read of that path (`startPlayback()` line ~454) is unchanged.

Refs: #274 (initial split), #315 (rename), #353 (introduced
coupling), #392 (partial cleanup, kept residuals), #408/#516/#520/#524/#526
(fullscreen + QuickTime fixes — fullscreen=true preserved as general UX).
